fixed image and rating for reviewbox, swapped match with new movies route
