+++ pluvio-app/src/components/Card.js
@@ -2,7 +2,6 @@ import styled from "styled-components";
 
 const Card = styled.div`
     background-color: #0b0725;
-    width: fit-content;
     height: fit-content;
     border-radius: 25px;
     color: white;
