+++ pluvio-app/src/components/ReviewBoxComponent/ReviewBox.css
@@ -8,7 +8,7 @@
 .view
 {
     display: flex;
-
+    justify-content: space-between;
 }
 
 .side
@@ -18,6 +18,7 @@
     margin: 0;
     justify-content: center;
     align-items: center;
+    gap: 20px;
 }
 
 @media screen and (max-width: 640px) {
